+++ CHANGELOG.md
@@ -69,7 +69,7 @@ different versioning scheme, following the Haskell community's
 * Fixed a race condition during `bond::ext::gRPC::unary_call` destruction.
 * Fixed the broken move constructor of `bond::bonded<T, Reader&>`.
 * Fixed the move constructor of `bond::value` to actually move the underlying reader.
-* Added `bond::blob_own` helper function that will return a `bond::blob` with
+* Added `bond::blob_prolong` helper function that will return a `bond::blob` with
   a copied data if the original one does not own the memory.
 * The `bond::OutputBuffer::GetBuffers` now can accept arbitrary STL-like containers.